requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
drf-yasg==1.21.7
pytest-django==4.7.0
pytest==8.0.1
factory-boy==3.3.0
Pillow==10.2.0
python-dateutil==2.8.2